Added player Economy System - Components are stack after each run - If the player buys an upgrade, he loose components - If  he doesn't have enough money, he can't buy the upgrade
Added "New Record" FX on the GameOver screen
- Jiggle effect on the text

+++ Assets/Components/PlayerStats/PlayerStat.cs
@@ -33,7 +33,7 @@ namespace Assets.Components.PlayerStats
 			return LstCoefPerLevel[playerLevel - 1];
 		}
 
-		public float GetUpgradeCost(int playerLevel)
+		public int GetUpgradeCost(int playerLevel)
 		{
 			if (LstCostPerLevel.Count < playerLevel)
 			{
